+++ CHANGELOG.md
@@ -38,6 +38,8 @@
 * Separate license key entries
 * Specify a “Text Domain” and “Domain Path”
 * Tested up to 3.9.0
+* Update Aihrus integration
+* Update copyright
 * Update EDD_SL_Plugin_Updater 1.1
 * Update `version_check` for free requirements
 * Use $plugin_assets than $plugin_path